requirements.txt
Flask==2.0.1
opencv-python-headless==4.5.3.56
pytesseract==0.3.8
langdetect==1.0.9
pillow==8.2.0
numpy==1.19.5
